refactor: move Home planes to a hook
